Adds a JVM runtime API
My original motivation here was to be able to easily inspect the command
line flags passed to the JVM for tuning our production configuration. I
went ahead and serialized the rest of the RuntimeMxBean while I was in
there since classpath and the like can also be useful.

I'd like to eventually make some nice UI around this, but I've been
unable to get miskweb to work locally. However, it's still a useful
first step to have an API.